Improve personality
Chatteroo responds to commands with a better personality.

The file path was also changed to remove the seedu sub-directory.
