+++ data/chatteroo.txt
@@ -1,4 +1,3 @@
-E | 0 | todo  | 2021-02-02 1900 | 2021-02-02 2000
-T | 0 | home
+E | 1 | todo  | 2021-02-02 1900 | 2021-02-02 2000
+T | 1 | home
 E | 0 | group meeting for 2103  | 2024-01-02 1800 | 2024-01-02 2000
-T | 0 | go
