+++ src/main/java/chatteroo/commands/DeleteCommand.java
@@ -1,0 +1,33 @@
+package chatteroo.commands;
+
+import chatteroo.tasks.TaskList;
+import chatteroo.ui.Ui;
+import chatteroo.storage.Storage;
+import chatteroo.tasks.Task;
+
+/**
+ * Deletes the specified task from the list of tasks.
+ */
+public class DeleteCommand extends Command {
+    private int taskNum;
+
+    /**
+     * Constructor for the DeleteCommand class.
+     * @param taskNum The task number.
+     */
+    public DeleteCommand(int taskNum) {
+        super();
+        this.taskNum = taskNum;
+    }
+
+    /**
+     * @inheritDoc
+     */
+    @Override
+    public String execute(TaskList tasks, Ui ui, Storage storage) {
+        Task deletedTask = tasks.getTask(taskNum);
+        tasks.deleteTask(taskNum);
+        int listCount = tasks.getTaskListSize();
+        return ui.showDeleteTaskResponse(deletedTask, listCount);
+    }
+}
