+++ src/main/java/chatteroo/commands/ExitCommand.java
@@ -1,0 +1,28 @@
+package chatteroo.commands;
+
+import chatteroo.tasks.TaskList;
+import chatteroo.ui.Ui;
+import chatteroo.storage.Storage;
+
+/**
+ * Prints the exit message.
+ */
+public class ExitCommand extends Command {
+
+    /**
+     * Constructor for the ExitCommand class.
+     */
+    public ExitCommand() {
+        super();
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public String execute(TaskList tasks, Ui ui, Storage storage) {
+        return ui.showByeResponse();
+    }
+
+
+}
