+++ src/main/java/chatteroo/commands/MarkCommand.java
@@ -1,0 +1,30 @@
+package chatteroo.commands;
+
+import chatteroo.tasks.TaskList;
+import chatteroo.ui.Ui;
+import chatteroo.storage.Storage;
+
+/**
+ * Marks the specified task as done.
+ */
+public class MarkCommand extends Command {
+    private int taskNum;
+
+    /**
+     * Constructor for the MarkCommand class.
+     * @param taskNum The task number.
+     */
+    public MarkCommand(int taskNum) {
+        super();
+        this.taskNum = taskNum;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public String execute(TaskList tasks, Ui ui, Storage storage) {
+        tasks.markTaskAsDone(taskNum);
+        return ui.showMarkDoneTaskResponse() + tasks.getTask(taskNum).toString();
+    }
+}
