+++ src/main/java/chatteroo/ui/MainWindow.java
@@ -1,0 +1,62 @@
+package chatteroo.ui;
+
+import javafx.fxml.FXML;
+import javafx.scene.control.Button;
+import javafx.scene.control.ScrollPane;
+import javafx.scene.control.TextField;
+import javafx.scene.image.Image;
+import javafx.scene.layout.AnchorPane;
+import javafx.scene.layout.VBox;
+
+import chatteroo.Chatteroo;
+
+import java.io.IOException;
+
+/**
+ * Controller for MainWindow. Provides the layout for the other controls.
+ */
+public class MainWindow extends AnchorPane {
+    @FXML
+    private ScrollPane scrollPane;
+    @FXML
+    private VBox dialogContainer;
+    @FXML
+    private TextField userInput;
+    @FXML
+    private Button sendButton;
+
+    private Chatteroo chatteroo;
+
+    private Image dukeImage = new Image(this.getClass().getResourceAsStream("/images/buffkangaroo.jpg"));
+    private Image userImage = new Image(this.getClass().getResourceAsStream("/images/minikangaroo.jpg"));
+
+    @FXML
+    public void initialize() {
+        scrollPane.vvalueProperty().bind(dialogContainer.heightProperty());
+
+        String welcomeText = Ui.showWelcomeResponse();
+        dialogContainer.getChildren().addAll(
+                DialogBox.getDukeDialog(welcomeText, dukeImage)
+        );
+    }
+
+    public void setChatteroo(Chatteroo d) {
+        chatteroo = d;
+    }
+
+    /**
+     * Creates two dialog boxes, one echoing user input and the other containing Duke's reply and then appends them to
+     * the dialog container. Clears the user input after processing.
+     */
+    @FXML
+    private void handleUserInput() throws IOException {
+        String input = userInput.getText();
+        String response = chatteroo.run(input);
+        dialogContainer.getChildren().addAll(
+                DialogBox.getUserDialog(input, userImage),
+                DialogBox.getDukeDialog(response, dukeImage)
+        );
+        userInput.clear();
+    }
+}
+
